+++ target-root/usr/local/lib/solidgroundux/common/sgnd-bootstrap.sh
@@ -3,8 +3,8 @@
 # -------------------------------------------------------------------------------------
 # Metadata:
 #   Version     : 1.8
-#   Build       : 2621201
-#   Checksum    : 6f12e4508d97781c3f416568b1eef13ceb91817bd315a02e8a07835f02db4040
+#   Build       : 2622101
+#   Checksum    : 61c5524041f5bba4126a3455321bb61461e6f9d427323015f4362795ab655d73
 #   Source      : sgnd-bootstrap.sh
 #   Type        : library
 #   Group       : Bootstrap