+++ target-root/usr/local/libexec/solidgroundux/console-modules/20-active-directory.sh
@@ -3,8 +3,8 @@
 # ----------------------------------------------------------------------------------
 # Metadata:
 #   Version     : 1.8
-#   Build       : 2621822
-#   Checksum    : 4bbd4e33d294e03bf2e00e0a6138edd45884c58114098372a75b7ce0be7883e8
+#   Build       : 2622101
+#   Checksum    : f7b86f86053cc2bc6b530f86dc1911b5211ceca3d603cb3bec0665da93126e1e
 #   Source      : 20-active-directory.sh
 #   Type        : module
 #   Group       : SolidGround Console