+++ target-root/usr/local/libexec/solidgroundux/console-modules/26-storage.sh
@@ -3,8 +3,8 @@
 # ----------------------------------------------------------------------------------
 # Metadata:
 #   Version     : 1.8
-#   Build       : 2621822
-#   Checksum    : be634d463666b522e62f80d2ed2b55f5023bb07944d4b3dbfa80e4a5c13ac19a
+#   Build       : 2622101
+#   Checksum    : 2b101c8eed4cfa64c40d0a8e3955115e2fb7b0b7c726fb4b55f39182389b85d5
 #   Source      : 26-storage-v4.sh
 #   Type        : module
 #   Group       : SolidGround Console