+++ target-root/usr/local/libexec/solidgroundux/deploy-workspace.sh
@@ -4,8 +4,8 @@
 # -------------------------------------------------------------------------------------
 # Metadata:
 #   Version     : 1.8
-#   Build       : 2621822
-#   Checksum    : 0edf29e11cb3679fe455ead45009b8f4364b76429b0cf39e48527ef6f7b57afb
+#   Build       : 2622101
+#   Checksum    : fea0b3d8b5aba392ed47945acf97e7c0579feb8ee780207bb6387b2175427339
 #   Source      : deploy-workspace-v4.sh
 #   Type        : script
 #   Group       : SDK Tools